+++ packages/generator-adp/src/telemetryEvents/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Event names for telemetry for the adaptation project fiori launchpad configuration generator
+ */
+export enum EventName {
+    ADAPTATION_PROJECT_CREATED = 'ADAPTATION_PROJECT_CREATED'
+}