+++ packages/preview-middleware/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @sap-ux/preview-middleware
 
+## 0.19.5
+
+### Patch Changes
+
+-   6cedb61: fix: enable telemetry for quickactions in adp
+
+## 0.19.4
+
+### Patch Changes
+
+-   Updated dependencies [ced5edf]
+    -   @sap-ux/adp-tooling@0.13.21
+    -   @sap-ux/system-access@0.5.34
+
+## 0.19.3
+
+### Patch Changes
+
+-   @sap-ux/adp-tooling@0.13.20
+
 ## 0.19.2
 
 ### Patch Changes